+++ omnijson/omni_gettransaction.go
@@ -32,12 +32,12 @@ type OmniGettransactionResult struct {
 	Amount          string `json:"amount"`
 	BlockHash       string `json:"blockhash"`
 	InvalidReason   string `json:"invalidreason"`
-	Version         int32  `json:"version"`
-	TypeInt         int32  `json:"type_int"`
-	PropertyID      int32  `json:"propertyid"`
-	BlockTimestamp  int32  `json:"blocktime"`
-	PositionInBlock int32  `json:"positioninblock"`
-	BlockHeight     int32  `json:"block"`
+	Version         int64  `json:"version"`
+	TypeInt         int64  `json:"type_int"`
+	PropertyID      int64  `json:"propertyid"`
+	BlockTimestamp  int64  `json:"blocktime"`
+	PositionInBlock int64  `json:"positioninblock"`
+	BlockHeight     int64  `json:"block"`
 	Confirmations   uint32 `json:"confirmations"`
 	Mine            bool   `json:"ismine"`
 	Divisible       bool   `json:"divisible"`
